+++ .github/workflows/bench.yml
@@ -1,5 +1,8 @@
 name: Benchmarks
-on: [push, pull_request]
+on:
+  push:
+    branches:
+      - main
 
 permissions:
   contents: write
@@ -16,17 +19,17 @@ jobs:
           toolchain: 1.56.1
           override: true
       - name: Run benchmark
-        run: cargo bench fft
-      # - name: Store benchmark result
-      #   uses: benchmark-action/github-action-benchmark@v1
-      #   with:
-      #     name: halo2 Benchmark
-      #     tool: 'cargo'
-      #     output-file-path: output.txt
-      #     github-token: ${{ secrets.GITHUB_TOKEN }}
-      #     auto-push: true
-      #     # Show alert with commit comment on detecting possible performance regression
-      #     alert-threshold: '200%'
-      #     comment-on-alert: true
-      #     fail-on-alert: true
-      #     alert-comment-cc-users: '@str4d'
+        run: cargo bench -- --output-format bencher | tee output.txt
+      - name: Store benchmark result
+        uses: benchmark-action/github-action-benchmark@v1
+        with:
+          name: halo2 Benchmark
+          tool: 'cargo'
+          output-file-path: output.txt
+          github-token: ${{ secrets.GITHUB_TOKEN }}
+          auto-push: true
+          # Show alert with commit comment on detecting possible performance regression
+          alert-threshold: '200%'
+          comment-on-alert: true
+          fail-on-alert: true
+          alert-comment-cc-users: '@str4d'
