+++ .github/workflows/ci.yml
@@ -1,146 +1,146 @@
-# name: CI checks
+name: CI checks
 
-# on: [push, pull_request]
+on: [push, pull_request]
 
-# jobs:
-#   test:
-#     name: Test on ${{ matrix.os }}
-#     runs-on: ${{ matrix.os }}
-#     strategy:
-#       matrix:
-#         os: [ubuntu-latest, windows-latest, macOS-latest]
+jobs:
+  test:
+    name: Test on ${{ matrix.os }}
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest, windows-latest, macOS-latest]
 
-#     steps:
-#       - uses: actions/checkout@v3
-#       - uses: actions-rs/toolchain@v1
-#         with:
-#           toolchain: 1.56.1
-#           override: true
-#       - name: Run tests
-#         uses: actions-rs/cargo@v1
-#         with:
-#           command: test
-#           args: --verbose --release --all --all-features
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions-rs/toolchain@v1
+        with:
+          toolchain: 1.56.1
+          override: true
+      - name: Run tests
+        uses: actions-rs/cargo@v1
+        with:
+          command: test
+          args: --verbose --release --all --all-features
 
-#   build:
-#     name: Build target ${{ matrix.target }}
-#     runs-on: ubuntu-latest
-#     strategy:
-#       matrix:
-#         target:
-#           - wasm32-unknown-unknown
-#           - wasm32-wasi
+  build:
+    name: Build target ${{ matrix.target }}
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        target:
+          - wasm32-unknown-unknown
+          - wasm32-wasi
 
-#     steps:
-#       - uses: actions/checkout@v3
-#       - uses: actions-rs/toolchain@v1
-#         with:
-#           toolchain: 1.56.1
-#           override: true
-#       - name: Add target
-#         run: rustup target add ${{ matrix.target }}
-#       - name: cargo build
-#         uses: actions-rs/cargo@v1
-#         with:
-#           command: build
-#           args: --features dev-graph,gadget-traces,unstable --target ${{ matrix.target }}
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions-rs/toolchain@v1
+        with:
+          toolchain: 1.56.1
+          override: true
+      - name: Add target
+        run: rustup target add ${{ matrix.target }}
+      - name: cargo build
+        uses: actions-rs/cargo@v1
+        with:
+          command: build
+          args: --features dev-graph,gadget-traces,unstable --target ${{ matrix.target }}
 
-#   bitrot:
-#     name: Bitrot check
-#     runs-on: ubuntu-latest
+  bitrot:
+    name: Bitrot check
+    runs-on: ubuntu-latest
 
-#     steps:
-#       - uses: actions/checkout@v3
-#       - uses: actions-rs/toolchain@v1
-#         with:
-#           toolchain: 1.56.1
-#           override: true
-#       # Build benchmarks to prevent bitrot
-#       - name: Build benchmarks
-#         uses: actions-rs/cargo@v1
-#         with:
-#           command: build
-#           args: --benches --examples --all-features
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions-rs/toolchain@v1
+        with:
+          toolchain: 1.56.1
+          override: true
+      # Build benchmarks to prevent bitrot
+      - name: Build benchmarks
+        uses: actions-rs/cargo@v1
+        with:
+          command: build
+          args: --benches --examples --all-features
 
-#   book:
-#     name: Book tests
-#     runs-on: ubuntu-latest
-#     steps:
-#       - uses: actions/checkout@v3
-#       - uses: actions-rs/toolchain@v1
-#         with:
-#           toolchain: 1.56.1
-#           override: true
-#       - name: cargo build
-#         uses: actions-rs/cargo@v1
-#         with:
-#           command: build
-#       - name: Setup mdBook
-#         uses: peaceiris/actions-mdbook@v1
-#         with:
-#           mdbook-version: '0.4.5'
-#       - name: Test halo2 book
-#         run: mdbook test -L target/debug/deps book/
+  book:
+    name: Book tests
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions-rs/toolchain@v1
+        with:
+          toolchain: 1.56.1
+          override: true
+      - name: cargo build
+        uses: actions-rs/cargo@v1
+        with:
+          command: build
+      - name: Setup mdBook
+        uses: peaceiris/actions-mdbook@v1
+        with:
+          mdbook-version: '0.4.5'
+      - name: Test halo2 book
+        run: mdbook test -L target/debug/deps book/
 
-#   codecov:
-#     name: Code coverage
-#     runs-on: ubuntu-latest
+  codecov:
+    name: Code coverage
+    runs-on: ubuntu-latest
 
-#     steps:
-#       - uses: actions/checkout@v3
-#       # Use stable for this to ensure that cargo-tarpaulin can be built.
-#       - uses: actions-rs/toolchain@v1
-#         with:
-#           toolchain: stable
-#           override: true
-#       - name: Install cargo-tarpaulin
-#         uses: actions-rs/cargo@v1
-#         with:
-#           command: install
-#           args: cargo-tarpaulin
-#       - name: Generate coverage report
-#         uses: actions-rs/cargo@v1
-#         with:
-#           command: tarpaulin
-#           args: --all-features --timeout 600 --out Xml
-#       - name: Upload coverage to Codecov
-#         uses: codecov/codecov-action@v3.1.0
+    steps:
+      - uses: actions/checkout@v3
+      # Use stable for this to ensure that cargo-tarpaulin can be built.
+      - uses: actions-rs/toolchain@v1
+        with:
+          toolchain: stable
+          override: true
+      - name: Install cargo-tarpaulin
+        uses: actions-rs/cargo@v1
+        with:
+          command: install
+          args: cargo-tarpaulin
+      - name: Generate coverage report
+        uses: actions-rs/cargo@v1
+        with:
+          command: tarpaulin
+          args: --all-features --timeout 600 --out Xml
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v3.1.0
 
-#   doc-links:
-#     name: Intra-doc links
-#     runs-on: ubuntu-latest
+  doc-links:
+    name: Intra-doc links
+    runs-on: ubuntu-latest
 
-#     steps:
-#       - uses: actions/checkout@v3
-#       - uses: actions-rs/toolchain@v1
-#         with:
-#           toolchain: 1.56.1
-#           override: true
-#       - name: cargo fetch
-#         uses: actions-rs/cargo@v1
-#         with:
-#           command: fetch
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions-rs/toolchain@v1
+        with:
+          toolchain: 1.56.1
+          override: true
+      - name: cargo fetch
+        uses: actions-rs/cargo@v1
+        with:
+          command: fetch
 
-#       # Ensure intra-documentation links all resolve correctly
-#       # Requires #![deny(intra_doc_link_resolution_failure)] in crates.
-#       - name: Check intra-doc links
-#         uses: actions-rs/cargo@v1
-#         with:
-#           command: doc
-#           args: --all --document-private-items
+      # Ensure intra-documentation links all resolve correctly
+      # Requires #![deny(intra_doc_link_resolution_failure)] in crates.
+      - name: Check intra-doc links
+        uses: actions-rs/cargo@v1
+        with:
+          command: doc
+          args: --all --document-private-items
 
-#   fmt:
-#     name: Rustfmt
-#     timeout-minutes: 30
-#     runs-on: ubuntu-latest
-#     steps:
-#       - uses: actions/checkout@v3
-#       - uses: actions-rs/toolchain@v1
-#         with:
-#           toolchain: 1.56.1
-#           override: true
-#       - run: rustup component add rustfmt
-#       - uses: actions-rs/cargo@v1
-#         with:
-#           command: fmt
-#           args: --all -- --check
+  fmt:
+    name: Rustfmt
+    timeout-minutes: 30
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions-rs/toolchain@v1
+        with:
+          toolchain: 1.56.1
+          override: true
+      - run: rustup component add rustfmt
+      - uses: actions-rs/cargo@v1
+        with:
+          command: fmt
+          args: --all -- --check
