+++ .github/workflows/lints-beta.yml
@@ -1,27 +1,27 @@
-# name: Beta lints
+name: Beta lints
 
-# # These lints are only informative, so we only run them directly on branches
-# # and not trial-merges of PRs, to reduce noise.
-# on: push
+# These lints are only informative, so we only run them directly on branches
+# and not trial-merges of PRs, to reduce noise.
+on: push
 
-# jobs:
-#   clippy-beta:
-#     name: Clippy (beta)
-#     timeout-minutes: 30
-#     runs-on: ubuntu-latest
-#     continue-on-error: true
+jobs:
+  clippy-beta:
+    name: Clippy (beta)
+    timeout-minutes: 30
+    runs-on: ubuntu-latest
+    continue-on-error: true
 
-#     steps:
-#       - uses: actions/checkout@v3
-#       - uses: actions-rs/toolchain@v1
-#         with:
-#           toolchain: beta
-#           components: clippy
-#           override: true
-#       - name: Run Clippy (beta)
-#         uses: actions-rs/clippy-check@v1
-#         continue-on-error: true
-#         with:
-#           name: Clippy (beta)
-#           token: ${{ secrets.GITHUB_TOKEN }}
-#           args: --all-features --all-targets -- -W clippy::all
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions-rs/toolchain@v1
+        with:
+          toolchain: beta
+          components: clippy
+          override: true
+      - name: Run Clippy (beta)
+        uses: actions-rs/clippy-check@v1
+        continue-on-error: true
+        with:
+          name: Clippy (beta)
+          token: ${{ secrets.GITHUB_TOKEN }}
+          args: --all-features --all-targets -- -W clippy::all
