+++ .github/workflows/lints-stable.yml
@@ -1,24 +1,24 @@
-# name: Stable lints
+name: Stable lints
 
-# # We only run these lints on trial-merges of PRs to reduce noise.
-# on: pull_request
+# We only run these lints on trial-merges of PRs to reduce noise.
+on: pull_request
 
-# jobs:
-#   clippy:
-#     name: Clippy (1.56.1)
-#     timeout-minutes: 30
-#     runs-on: ubuntu-latest
+jobs:
+  clippy:
+    name: Clippy (1.56.1)
+    timeout-minutes: 30
+    runs-on: ubuntu-latest
 
-#     steps:
-#       - uses: actions/checkout@v3
-#       - uses: actions-rs/toolchain@v1
-#         with:
-#           toolchain: 1.56.1
-#           components: clippy
-#           override: true
-#       - name: Run clippy
-#         uses: actions-rs/clippy-check@v1
-#         with:
-#           name: Clippy (1.56.1)
-#           token: ${{ secrets.GITHUB_TOKEN }}
-#           args: --all-features --all-targets -- -D warnings
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions-rs/toolchain@v1
+        with:
+          toolchain: 1.56.1
+          components: clippy
+          override: true
+      - name: Run clippy
+        uses: actions-rs/clippy-check@v1
+        with:
+          name: Clippy (1.56.1)
+          token: ${{ secrets.GITHUB_TOKEN }}
+          args: --all-features --all-targets -- -D warnings
